+++ packages/ckb-sdk-core/src/index.ts
@@ -1,13 +1,17 @@
 import RPC from '@nervosnetwork/ckb-sdk-rpc'
-import Address from '@nervosnetwork/ckb-sdk-address'
+import { ArgumentRequired } from '@nervosnetwork/ckb-sdk-utils/lib/exceptions'
 import * as utils from '@nervosnetwork/ckb-sdk-utils'
 
-interface DepCellInfo {
-  hashType: CKBComponents.ScriptHashType
-  codeHash: CKBComponents.Hash256
-  outPoint: CKBComponents.OutPoint
-}
+import generateRawTransaction from './generateRawTransaction'
+
+import Address from './address'
+import loadCells from './loadCells'
+
+const hrpSize = 6
+
 class Core {
+  public cells: Map<string, CachedCell[]> = new Map()
+
   public rpc: RPC
 
   public utils = utils
@@ -81,6 +85,21 @@ class Core {
       codeHashIndex,
     })
 
+  public generateLockHash = (
+    publicKeyHash: string,
+    deps: Omit<DepCellInfo, 'outPoint'> | undefined = this.config.secp256k1Dep
+  ) => {
+    if (!deps) {
+      throw new ArgumentRequired('deps')
+    }
+
+    return this.utils.scriptToHash({
+      hashType: deps.hashType,
+      codeHash: deps.codeHash,
+      args: publicKeyHash,
+    })
+  }
+
   public loadSecp256k1Dep = async () => {
     /**
      * cell list
@@ -116,6 +135,26 @@ class Core {
     return this.config.secp256k1Dep
   }
 
+  public loadCells = async ({
+    lockHash,
+    start = 0,
+    end,
+    STEP = 100,
+    save = false,
+  }: {
+    lockHash: string
+    start?: string | number
+    end?: string | number
+    STEP?: number
+    save?: boolean
+  }) => {
+    const cells = await loadCells({ lockHash, start, end, STEP, rpc: this.rpc })
+    if (save) {
+      this.cells.set(lockHash, cells)
+    }
+    return cells
+  }
+
   public signWitnesses = (key: string | Address) => ({
     transactionHash,
     witnesses = [],
@@ -123,32 +162,28 @@ class Core {
     transactionHash: string
     witnesses: CKBComponents.Witness[]
   }) => {
-    if (!key) throw new Error('Private key or address object is required')
-    if (!transactionHash) throw new Error('Transaction hash is required')
+    if (!key) throw new ArgumentRequired('Private key or address object')
+    if (!transactionHash) throw new ArgumentRequired('Transaction hash')
 
     const addrObj = typeof key === 'string' ? this.generateAddress(key) : key
     const signedWitnesses = witnesses.map(witness => {
-      const oldData = witness.data || []
       const s = this.utils.blake2b(32, null, null, this.utils.PERSONAL)
+      const witnessBytes = this.utils.hexToBytes(witness)
       s.update(this.utils.hexToBytes(transactionHash))
-      oldData.forEach(datum => {
-        s.update(this.utils.hexToBytes(datum))
-      })
+      s.update(witnessBytes)
       const message = `0x${s.digest('hex')}`
-      const data = [addrObj.signRecoverable(message), ...oldData]
-      return {
-        data,
-      }
+      const data = [...this.utils.hexToBytes(addrObj.signRecoverable(message)), ...witnessBytes]
+      return this.utils.bytesToHex(new Uint8Array(data))
     })
     return signedWitnesses
   }
 
   public signTransaction = (key: string | Address) => (transaction: CKBComponents.RawTransaction) => {
-    if (!key) throw new Error('Private key or address object is required')
-    if (!transaction) throw new Error('Transaction is required')
-    if (!transaction.witnesses) throw new Error('Witnesses is required')
+    if (!key) throw new ArgumentRequired('Private key or address object')
+    if (!transaction) throw new ArgumentRequired('Transaction')
+    if (!transaction.witnesses) throw new ArgumentRequired('Witnesses')
     if (transaction.witnesses.length < transaction.inputs.length) throw new Error('Invalid count of witnesses')
-    if (!transaction.outputsData) throw new Error('OutputsData is required')
+    if (!transaction.outputsData) throw new ArgumentRequired('OutputsData')
     if (transaction.outputsData.length < transaction.outputs.length) throw new Error('Invalid count of outputsData')
 
     const transactionHash = this.utils.rawTransactionToHash(transaction)
@@ -160,6 +195,52 @@ class Core {
       ...transaction,
       witnesses: signedWitnesses,
     }
+  }
+
+  public generateRawTransaction = async ({
+    fromAddress,
+    toAddress,
+    capacity,
+    safeMode = true,
+    cells = [],
+    deps = this.config.secp256k1Dep!,
+  }: {
+    fromAddress: string
+    toAddress: string
+    capacity: string | bigint
+    safeMode: boolean
+    cells: CachedCell[]
+    deps: DepCellInfo
+  }) => {
+    const [fromPublicKeyHash, toPublicKeyHash] = [fromAddress, toAddress].map((addr: string) => {
+      const addressPayload = this.utils.parseAddress(addr, 'hex')
+      return `0x${addressPayload.slice(hrpSize)}`
+    })
+
+    let availableCells = cells
+    if (!availableCells.length && deps) {
+      const lockHash = this.utils.scriptToHash({
+        codeHash: deps.codeHash,
+        hashType: deps.hashType,
+        args: fromPublicKeyHash,
+      })
+      const cachedCells = this.cells.get(lockHash)
+      if (cachedCells && cachedCells.length) {
+        availableCells = cachedCells
+      } else {
+        const fetchedCells = await this.loadCells({ lockHash, save: true })
+        availableCells = fetchedCells
+      }
+    }
+
+    return generateRawTransaction({
+      fromPublicKeyHash,
+      toPublicKeyHash,
+      capacity,
+      safeMode,
+      cells: availableCells,
+      deps,
+    })
   }
 }