+++ packages/ckb-types/index.d.ts
@@ -55,14 +55,19 @@ declare namespace CKBComponents {
   /* eslint-disable max-len */
   /**
    * @typedef Script, lock or type script
-   * @description Script, the script model in CKB. CKB scripts use UNIX standard execution environment. Each script binary should contain a main function with the following signature `int main(int argc, char* argv[]);`. CKB will concat `signed_args` and `args`, then use the concatenated array to fill `argc/argv` part, then start the script execution. Upon termination, the executed `main` function here will provide a return code, `0` means the script execution succeeds, other values mean the execution fails.
+   * @description Script, the script model in CKB. CKB scripts use UNIX standard execution environment.
+   *              Each script binary should contain a main function with the following signature `int main(int argc, char* argv[]);`.
+   *              CKB will concat  `args`, then use the concatenated array to fill `argc/argv` part, then start the script execution.
+   *              Upon termination, the executed `main` function here will provide a return code,
+   *              `0` means the script execution succeeds, other values mean the execution fails.
    * @property args, arguments.
-   * @property codeHash, point to its dependency, if the referred dependency is listed in the deps field in a transaction, the codeHash means the hash of the referred cell's data.
+   * @property codeHash, point to its dependency, if the referred dependency is listed in the deps field in a transaction,
+   *                     the codeHash means the hash of the referred cell's data.
    * @property hashType, a enumerate indicates the type of the code which is referened by the code hash
    */
   /* eslint-enable max-len */
   export interface Script {
-    args: Bytes[]
+    args: Bytes
     codeHash: Hash256
     hashType: ScriptHashType
   }
@@ -110,9 +115,7 @@ declare namespace CKBComponents {
     depType: DepType
   }
 
-  export interface Witness {
-    data: Hash[]
-  }
+  export type Witness = Bytes
 
   /**
    * @typedef RawTransaction, raw transaction object
@@ -177,8 +180,8 @@ declare namespace CKBComponents {
 
   /**
    * @typedef BlockHeader, header of a block
+   * @property compactTarget
    * @property dao
-   * @property difficulty
    * @property epoch
    * @property hash
    * @property number
@@ -187,14 +190,12 @@ declare namespace CKBComponents {
    * @property nonce
    * @property timestamp
    * @property transactionsRoot
-   * @property unclesCount
    * @property unclesHash
-   * @property witnessesRoot
    * @property version
    */
   export interface BlockHeader {
+    compactTarget: Hash
     dao: DAO
-    difficulty: Difficulty
     epoch: EpochInHeader
     hash: Hash256
     number: BlockNumber
@@ -203,9 +204,7 @@ declare namespace CKBComponents {
     nonce: Nonce
     timestamp: Timestamp
     transactionsRoot: Hash256
-    unclesCount: Count
     unclesHash: Hash256
-    witnessesRoot: Hash256
     version: Version
   }
 
@@ -263,6 +262,7 @@ declare namespace CKBComponents {
    */
 
   export interface CellIncludingOutPoint {
+    blockHash: Hash256
     capacity: Capacity
     lock: Script
     outPoint: OutPoint | null
@@ -326,7 +326,7 @@ declare namespace CKBComponents {
   }
 
   export interface Epoch {
-    difficulty: String
+    compactTarget: Hash
     length: String
     number: String
     startNumber: String