+++ packages/ckb-sdk-utils/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.22.0](https://github.com/nervosnetwork/ckb-sdk-js/compare/v0.21.1...v0.22.0) (2019-10-05)
+
+
+### Features
+
+* **core:** move the address module into the core module ([10dd017](https://github.com/nervosnetwork/ckb-sdk-js/commit/10dd017))
+* **type:** update the type of args ([09d649a](https://github.com/nervosnetwork/ckb-sdk-js/commit/09d649a))
+* **type:** update the type of witness ([71f53b0](https://github.com/nervosnetwork/ckb-sdk-js/commit/71f53b0))
+* **utils:** update the interface of utils.parseAddress ([90feb91](https://github.com/nervosnetwork/ckb-sdk-js/commit/90feb91))
+
+
+### BREAKING CHANGES
+
+* **type:** change the type of args from string[] to string
+* **type:** change the type of witnes from { data: string[] } to string
+* **utils:** remove the prefix from the parameter list of utils.parseAddress
+* **core:** move the address module into the core module
+
+
+
+
+
 ## [0.21.1](https://github.com/nervosnetwork/ckb-sdk-js/compare/v0.21.0...v0.21.1) (2019-09-24)
 
 **Note:** Version bump only for package @nervosnetwork/ckb-sdk-utils